Crate solana_transaction_status

Re-exports

pub use crate::extract_memos::extract_and_fmt_memos;

Modules

extract_memos

option_serializer

parse_accounts

parse_address_lookup_table

parse_associated_token

parse_bpf_loader

parse_instruction

parse_stake

parse_system

parse_token

parse_vote

token_balances

Structs

BlockEncodingOptions

ConfirmedBlock

ConfirmedTransactionStatusWithSignature

ConfirmedTransactionWithStatusMeta

EncodedConfirmedBlock

EncodedConfirmedTransactionWithStatusMeta

EncodedTransactionWithStatusMeta

EntrySummary

InnerInstruction

InnerInstructions

Reward

RewardsAndNumPartitions

TransactionByAddrInfo

TransactionStatus

TransactionStatusMeta

TransactionTokenBalance

UiAccountsList

UiAddressTableLookup

A duplicate representation of a MessageAddressTableLookup, in raw format, for pretty JSON serialization

UiCompiledInstruction

A duplicate representation of a CompiledInstruction for pretty JSON serialization

UiConfirmedBlock

UiInnerInstructions

UiLoadedAddresses

A duplicate representation of LoadedAddresses

UiParsedMessage

A duplicate representation of a Message, in parsed format, for pretty JSON serialization

UiPartiallyDecodedInstruction

A partially decoded CompiledInstruction that includes explicit account addresses

UiRawMessage

A duplicate representation of a Message, in raw format, for pretty JSON serialization

UiTransaction

A duplicate representation of a Transaction for pretty JSON serialization

UiTransactionReturnData

UiTransactionStatusMeta

A duplicate representation of TransactionStatusMeta with err field

UiTransactionTokenBalance

VersionedConfirmedBlock

VersionedConfirmedBlockWithEntries

VersionedConfirmedTransactionWithStatusMeta

VersionedTransactionWithStatusMeta

Enums

ConvertBlockError

EncodeError

EncodedTransaction

RewardType

TransactionBinaryEncoding

TransactionConfirmationStatus

TransactionDetails

TransactionWithStatusMeta

UiInstruction

A duplicate representation of an Instruction for pretty JSON serialization

UiMessage

UiParsedInstruction

UiReturnDataEncoding

UiTransactionEncoding

Traits

Encodable

Represents types that can be encoded into one of several encoding formats

EncodableWithMeta

Represents types that can be encoded into one of several encoding formats

Functions

map_inner_instructions

Maps a list of inner instructions from solana_sdk into a list of this crate’s representation of inner instructions (with instruction indices).

parse_ui_inner_instructions

parse_ui_instruction

Type Aliases

Rewards