Crate solana_transaction_status

Re-exports

pub use crate::extract_memos::extract_and_fmt_memos;

Modules

extract_memos

option_serializer

parse_accounts

parse_address_lookup_table

parse_associated_token

parse_bpf_loader

parse_instruction

parse_stake

parse_system

parse_token

parse_vote

token_balances

Structs

BlockEncodingOptions

ConfirmedBlock

ConfirmedTransactionStatusWithSignature

ConfirmedTransactionWithStatusMeta

EncodedConfirmedBlock

EncodedConfirmedTransactionWithStatusMeta

EncodedTransactionWithStatusMeta

InnerInstruction

InnerInstructions

Reward

TransactionByAddrInfo

TransactionStatus

TransactionStatusMeta

TransactionTokenBalance

UiAccountsList

UiAddressTableLookup

A duplicate representation of a MessageAddressTableLookup, in raw format, for pretty JSON serialization

UiCompiledInstruction

A duplicate representation of a CompiledInstruction for pretty JSON serialization

UiConfirmedBlock

UiInnerInstructions

UiLoadedAddresses

A duplicate representation of LoadedAddresses

UiParsedMessage

A duplicate representation of a Message, in parsed format, for pretty JSON serialization

UiPartiallyDecodedInstruction

A partially decoded CompiledInstruction that includes explicit account addresses

UiRawMessage

A duplicate representation of a Message, in raw format, for pretty JSON serialization

UiTransaction

A duplicate representation of a Transaction for pretty JSON serialization

UiTransactionReturnData

UiTransactionStatusMeta

A duplicate representation of TransactionStatusMeta with err field

UiTransactionTokenBalance

VersionedConfirmedBlock

VersionedConfirmedTransactionWithStatusMeta

VersionedTransactionWithStatusMeta

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

Enums

ConvertBlockError

EncodeError

EncodedTransaction

RewardType

TransactionBinaryEncoding

TransactionConfirmationStatus

TransactionDetails

TransactionWithStatusMeta

UiInstruction

A duplicate representation of an Instruction for pretty JSON serialization

UiMessage

UiParsedInstruction

UiReturnDataEncoding

UiTransactionEncoding

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

Constants

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

Traits

Encodable

Represents types that can be encoded into one of several encoding formats

EncodableWithMeta

Represents types that can be encoded into one of several encoding formats

JsonAccounts ğŸ”’

Type Definitions

Rewards